move settings validator
