+++ src/shared/settings/validator.js
